requirements.txt

openai
langchain
faiss-cpu
pdfminer.six
python-docx
            